🚚 move: Correct import path for LoginForm component and add new LoginForm implementation
